Extract error types to separate module
